Uploaded SQL Questions and Answers
I have uploaded 5 pairs of files: question_1.sql contains the SQL code to answer question 1. The result of the query (and answer to the question) can be found in question_1.csv. This is the case for questions 1 through 5. The questions relate to the same data I analysed in Power BI.